feat(booking): validate barberman ownership on cancel booking (#4)

+++ src/domain/application/use-cases/booking/cancel-booking/cancel-booking.ts
@@ -1,11 +1,12 @@
 import { type Either, left, right } from "../../../../../core/logic/either";
 import type { BookingsRepository } from "../../../repositories/bookings-repository";
 import { BookingNotFoundError } from "../../_errors/booking-not-found-error";
+import { NotAllowedError } from "../../_errors/not-allowed-error";
 import type { CancelBookingDTO } from "./cancel-booking-dto";
 import type { CancelBookingResponse } from "./cancel-booking-response";
 
 type CancelBookingUseCaseResponse = Either<
-  BookingNotFoundError,
+  BookingNotFoundError | NotAllowedError,
   CancelBookingResponse
 >;
 
@@ -14,13 +15,16 @@ export class CancelBookingUseCase {
 
   async execute({
     bookingId,
+    barbermanId,
   }: CancelBookingDTO): Promise<CancelBookingUseCaseResponse> {
     const booking = await this.bookingsRepository.findById(bookingId);
 
     if (!booking) {
       return left(new BookingNotFoundError());
     }
-
+    if (booking.barbermanId.toString() !== barbermanId) {
+      return left(new NotAllowedError());
+    }
     await this.bookingsRepository.delete(booking);
 
     return right({
